% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_characters}
\alias{get_characters}
\title{Get characters}
\usage{
get_characters()
}
\description{
\code{get_characters} returns a tibble of all the characters from the
`character` endpoint of
\href{https://rickandmortyapi.com/documentation/#character}{The Rick and Morty API}.
This information includes:
\itemize{
  \item id
  \item name
  \item status
  \item species
  \item type
  \item gender
  \item image
  \item episode
  \item url
  \item created
  \item origin.name
  \item origin.url
  \item location.name
  \item location.url
}
}
\examples{
\dontrun{
get_characters()
}

}
