% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_episodes}
\alias{get_episodes}
\title{Get episodes}
\usage{
get_episodes()
}
\description{
\code{get_episodes} returns a tibble of all the episodes from the
`episode` endpoint of
\href{https://rickandmortyapi.com/documentation/#episode}{The Rick and Morty API}.
This information includes:
\itemize{
  \item id
  \item name
  \item air_date
  \item episode
  \item characters
  \item url
  \item created
}
}
\examples{
\dontrun{
get_episodes()
}

}
