% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_locations}
\alias{get_locations}
\title{Get locations}
\usage{
get_locations()
}
\description{
\code{get_locations} returns a tibble of all the locations from the
`location` endpoint of
\href{https://rickandmortyapi.com/documentation/#location}{The Rick and Morty API}.
This information includes:
\itemize{
  \item id
  \item name
  \item type
  \item dimension
  \item residents
  \item url
  \item created
}
}
\examples{
\dontrun{
get_locations()
}

}
